Move Hashable, CustomHashMap to separate files. WIP build fixing
