+++ fp4ml-main/build.sbt
@@ -7,7 +7,7 @@ ScalariformKeys.preferences := sharedCodeFmt
 
 addCompilerPlugin(scalaMacros)
 
-libraryDependencies :=
+libraryDependencies ++=
   fp4mlMainDeps ++
   testDeps
