+++ fp4ml-spark/build.sbt
@@ -7,7 +7,7 @@ ScalariformKeys.preferences := sharedCodeFmt
 
 addCompilerPlugin(scalaMacros)
 
-libraryDependencies := 
+libraryDependencies ++=
   fp4mlSparkDeps ++
   testDeps
